+++ src/activities/reader/EpubReaderActivity.cpp
@@ -1010,6 +1010,14 @@ void EpubReaderActivity::executeReaderQuickAction(CrossPointSettings::LONG_PRESS
     case CrossPointSettings::LONG_MENU_FILE_TRANSFER:
       openFileTransfer();
       break;
+    case CrossPointSettings::LONG_MENU_TOGGLE_TILT_PAGE_TURN:
+      if (halTiltSensor.isAvailable()) {
+        SETTINGS.tiltPageTurn = SETTINGS.tiltPageTurn == CrossPointSettings::TILT_OFF ? CrossPointSettings::TILT_NORMAL
+                                                                                      : CrossPointSettings::TILT_OFF;
+        SETTINGS.saveToFile();
+        halTiltSensor.clearPendingEvents();
+      }
+      break;
     case CrossPointSettings::LONG_MENU_OFF:
     default:
       break;
@@ -1052,6 +1060,9 @@ bool EpubReaderActivity::executeShortPowerButtonAction() {
       return true;
     case CrossPointSettings::SHORT_PWRBTN::FILE_TRANSFER:
       executeReaderQuickAction(CrossPointSettings::LONG_MENU_FILE_TRANSFER);
+      return true;
+    case CrossPointSettings::SHORT_PWRBTN::TOGGLE_TILT_PAGE_TURN:
+      executeReaderQuickAction(CrossPointSettings::LONG_MENU_TOGGLE_TILT_PAGE_TURN);
       return true;
     default:
       return false;
@@ -1112,6 +1123,9 @@ bool EpubReaderActivity::executeLongPowerButtonAction() {
       return true;
     case CrossPointSettings::SHORT_PWRBTN::FILE_TRANSFER:
       executeReaderQuickAction(CrossPointSettings::LONG_MENU_FILE_TRANSFER);
+      return true;
+    case CrossPointSettings::SHORT_PWRBTN::TOGGLE_TILT_PAGE_TURN:
+      executeReaderQuickAction(CrossPointSettings::LONG_MENU_TOGGLE_TILT_PAGE_TURN);
       return true;
     default:
       return false;